+++ CHANGELOG.md
@@ -2,6 +2,28 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+### [2.53.16](https://github.com/EightfoldAI/octuple/compare/v2.53.15...v2.53.16) (2025-07-23)
+
+### Bug Fixes
+
+- tab issue causing frontend error ([#1023](https://github.com/EightfoldAI/octuple/issues/1023)) ([4189315](https://github.com/EightfoldAI/octuple/commits/418931595f93ae2a69e7e7e7325afba8911cdc02))
+
+### [2.53.15](https://github.com/EightfoldAI/octuple/compare/v2.53.14...v2.53.15) (2025-07-22)
+
+### Bug Fixes
+
+- tab: added a prop to enable standard key navigation for tabs component ([#953](https://github.com/EightfoldAI/octuple/issues/953)) ([7934881](https://github.com/EightfoldAI/octuple/commits/79348811e884f7a10cabbc15d31f6c65e464a3a1))
+
+### [2.53.14](https://github.com/EightfoldAI/octuple/compare/v2.53.13...v2.53.14) (2025-07-21)
+
+### [2.53.13](https://github.com/EightfoldAI/octuple/compare/v2.53.12...v2.53.13) (2025-07-21)
+
+### Bug Fixes
+
+- panel: panel role not assigned ([#1018](https://github.com/EightfoldAI/octuple/issues/1018)) ([5cca827](https://github.com/EightfoldAI/octuple/commits/5cca8275ccea2721b4975e1940b829f96c40cc5c))
+- pre-commit hook update ([#1016](https://github.com/EightfoldAI/octuple/issues/1016)) ([25f141f](https://github.com/EightfoldAI/octuple/commits/25f141fe44f0a874b748a9457a92c0f3bc1c9b7a))
+- select input will have ariaLabel ([#1019](https://github.com/EightfoldAI/octuple/issues/1019)) ([2c724cd](https://github.com/EightfoldAI/octuple/commits/2c724cdc7ef2d638fce58738954760a0130df8e3))
+
 ### [2.53.12](https://github.com/EightfoldAI/octuple/compare/v2.53.11...v2.53.12) (2025-06-25)
 
 ### Bug Fixes